AIM-181 vlm model package function refactoring & test.
