+++ vlm_model/routers/send_feedback.py
@@ -9,7 +9,7 @@ import base64
 from pathlib import Path
 
 from vlm_model.schemas.feedback import FeedbackResponse
-from vlm_model.utils.video_processing import process_video
+from vlm_model.utils.processing_video import process_video
 from vlm_model.exceptions import VideoProcessingError, ImageEncodingError
 from vlm_model.config import FEEDBACK_DIR, UPLOAD_DIR
